vakantieplanner v1.1.0

+++ app/Http/Controllers/applicationCheckController.php
@@ -16,6 +16,20 @@ class applicationCheckController extends Controller
     // Sent data to the standard page with pagination and multiple search function as active users only
     public function index(Request $request){
       $getuserinfo = User::where('active',0)->get();
+      $currentuser = Auth::user();
+
+      $openApplications = $this->getQueryBuilder($request)->where('status_id', 1)->orderBy('created_at','desc');
+      $closedApplications = $this->getQueryBuilder($request)->where('status_id','!=', 1)->orderBy('created_at','desc');
+
+      if(Auth::user()->hasRole('employee')){
+        $openApplications = $openApplications->where('user_id',$currentuser->id);
+        $closedApplications = $closedApplications->where('user_id',$currentuser->id);
+      }
+
+      return view('/pages/applicationcheck/applicationcheck', ['openApplications'=>$openApplications->paginate(7), 'closedApplications'=>$closedApplications->paginate(7), 'data'=>$request]);
+    }
+
+    private function getQueryBuilder($request){
       $applicationdatas = Application::query();
       $applicationstatus = ApplicationStatus::pluck('name');
 
@@ -28,20 +42,8 @@ class applicationCheckController extends Controller
       $applicationdatas->whereHas('user', function ($query) {
         return $query->where('active', 1);
       });
-
-      $currentuser = Auth::user();
-
-      if(Auth::user()->hasRole('employer')){
-        $result = $applicationdatas->orderBy('status_id','ASC')->paginate(7);
-      }
-      else if(Auth::user()->hasRole('employee')){
-        $result = $applicationdatas->orderBy('status_id','ASC')->where('user_id',$currentuser->id)->paginate(7);
-      }
-
-
-      return view('/pages/applicationcheck/applicationcheck', ['applicationdatas'=>$result, 'data'=>$request]);
+      return $applicationdatas;
     }
-
 
     // Function to sent to right detials to the edit application page
     public function editapplication($get_id){
@@ -54,43 +56,38 @@ class applicationCheckController extends Controller
     // Function to edit the status with the 2 active buttons on the edit application page
     public function status(request $request, $id){
       $statusupdate = Application::find($id);
-      $statusupdate->status_id = $request->buttonstatus;
-      $statusupdate->application_status_remark = $request->formRemark;
-      $statusupdate->save();
-      app('App\Http\Controllers\mailController')->statusUpdate($id);
-
-      if($request->buttonstatus == '2'){
-        $newEvent = new Event;
-        $newEvent = Event::create([
-          'name' => 'Verlof '. Auth::user()->name,
-          'startDateTime' => Carbon::parse($statusupdate->date_from),
-          'endDateTime' => Carbon::parse($statusupdate->date_till),
-        ]);
-
-        $statusupdate->google_calander_id = $newEvent->id;
-        $statusupdate->save();
-
-      }else if($request->buttonstatus == '3'){
-        $deletecalanderid = Application::find($id);
-        $eventId = $deletecalanderid->google_calander_id;
-
-        if($eventId != null){
-          $event = Event::find($eventId);
-          $event->delete();
+      if($statusupdate->status_id == $request->buttonstatus){
+        return redirect()->route('applicationcheck');
+      }else{
+        $statusupdate->status_id = $request->buttonstatus;
+        if($request->formRemark != null){
+          $statusupdate->application_status_remark = $request->formRemark;
         }
+        $statusupdate->save();
+        app('App\Http\Controllers\mailController')->statusUpdate($id);
+
+        if($request->buttonstatus == '2'){
+          $newEvent = new Event;
+          $newEvent = Event::create([
+            'name' => $statusupdate->user->name . ' vrij',
+            'startDateTime' => Carbon::parse($statusupdate->date_from),
+            'endDateTime' => Carbon::parse($statusupdate->date_till),
+          ]);
+
+          $statusupdate->google_calander_id = $newEvent->id;
+          $statusupdate->save();
+
+        }else if($request->buttonstatus == '3'){
+          $deletecalanderid = Application::find($id);
+          $eventId = $deletecalanderid->google_calander_id;
+
+          if($eventId != null){
+            $event = Event::find($eventId);
+            $event->delete();
+          }
+        }
+        return redirect()->route('applicationcheck');
 
       }
-
-      return redirect()->route('applicationcheckedit',$id );
-    }
-
-
-    // Function to set a remark as employer
-    public function pushremark(request $request, $applid){
-      $updateremark = Application::find($applid);
-      $updateremark->application_status_remark = $request->formRemark;
-      $updateremark->save();
-      return redirect()->route('applicationcheckedit',$applid );
-
     }
 }
